Fixed bug where ECALL is always labeled as EBREAK
This bug occured because of a simple mistake of mine, I treated
ECALL and EBREAK as though they had different opcodes when in reality
they had the same opcode with a different immediate. When creating the
dictionary, I used the same key for both functions with EBREAK after
the ECALL so it was overrided.